apps/api/pyproject.toml
[project]
name = "timbermap-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "google-cloud-storage (>=3.10.0,<4.0.0)",
    "google-cloud-secret-manager (>=2.26.0,<3.0.0)",
    "google-cloud-tasks (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "svix (>=1.89.0,<2.0.0)",
    "geoserver-rest (>=2.10.0,<3.0.0)",
    "resend (>=2.26.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
